Add the ability to save and load descriptor to the disk
Each block is saved in a classic netcf (v3) file with 64-bit offsets.
The files contains variables for each labels names, values; the values of
the block, and the samples & values of gradients.

Then multiple blocks files are grouped together in a single zip file
as well as another netcdf file containing the sparse labels (this is
similar to how .npz file work).